tci6614: add driver for cpintd pulse-level convertor
TCI6614 SOC has CP_INTD (a.k.a Common Platform Interrupt Distributor for
mapping some of the System interrupts to CP_INTC. This is the driver
for CP_INTD. This also extend the davinci_soc_info structure so that
individual SOC can define if there is support for CP_INTD or not. This code
is designed similar to the CP_INTC driver.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>