hwqueue: add hwqueue core infrastructure
This patch adds an hwqueue core framework roughly along the lines of the
hwspinlock architecture.  This framework is hardware/platform independent and
only provides an API abstraction on top of a registration/unregistration
mechanism for real hardware hwqueue implementations.

The abstraction elements in here are as follows:

 - "device" corresponds to a hardware queue manager.  In the case of Appleton
   there is only one instance of a hwqueue device in the system.  However,
   future devices may have more than one physical instance, for example,
   Keystone 2 devices have 2 instances of QMSS 1.5 devices.

 - "instance" corresponds to an actual hardware queue.  Instances are numbered
   per-device as offsets from the device's base_id.

- "handle" corresponds to a particular user of a hardware queue instance.
  Hardware queues may be opened by multiple entities in the kernel, and each
  entity may open the queue with a different read/write mode.

Supported functionality:

- Registration: The hwqueue core framework depends only on statically
  initialized state, and is therefore active very early in the kernel boot
  process.  This allows other drivers to use hwqueue APIs without worrying
  about actual probe order.  The registration/unregistration APIs allow
  underlying hardware implementations to add (and remove) themselves on the
  fly.

- File semantics: The hwqueue core framework enforces traditional Unix
  open() mode semantics on top of hardware queues.  These semantics include:
  * read/write modes: read-only, write-only, and read-write
  * creation mode: enforces that a queue not be already open
  * exclusive mode: enforces exclusive ownership on the queue

- Notification: Read-capable users of an hwqueue may register one or more
  notifiers on the queue.  These notifiers are called out independent of the
  underlying notification mechanism (accumulation, qpend interrupt, timer
  poll).  The hwqueue core framework provides a timer based poll
  implementation for hwqueue hardware that does not support interrupt driven
  notification.

- Debug visibility:  The hwqueue core framework provides reasonably detailed
  debugfs visibility into the queues managed through this framework.

This framework is nowhere near complete, and is sure to be extended quite a
bit more as we add real hardware drivers under this framework, and real API
users above it.  It would, however, be useful to get some early feedback on
the overall model, assumptions, etc. before we go too far down this road.

Signed-Off-by: Prabhu Kuttiyam <pkuttiyam@ti.com>
Signed-off-by: Cyril Chemparathy <cyril@ti.com>

hwqueue: cleanups, improve debugfs dump

hwqueue: avoid divide on queue push/pop

This patch speeds up hwqueue push and pop operations by eliminating a divide
operation in the processing.  Since the ARM CPU does not have a divide
instruction, this operation was spending measurable time in software divide
routines.

This divide was being done in mapping an hwqueue instance pointer to and from
a queue index.  With this patch in place, we round up the size of the hwqueue
instance data to the nearest power of two, and use faster shift operations
instead.

hwqueue: add branch optimization hints

This patch mainly adds branch optimization hints (likely/unlikely) for the
compiler's benefit.  In addition, we now enforce the presence of push and pop
handlers at device registration time instead of checking on every push/pop.
We've also removed a couple of impossible warning checks on helper routines
that convert back and forth between hwqueue handles and internal types.

hwqueue: reorder fields to group frequently used data

This patch reorders the contents of the hwqueue_device data structure, to
group together data that is frequently used in the data path.

hwqueue: simplify data structures, inline data path ops

This patch simplifies the hwqueue core data structures, by eliminating the
need for a distinct hwqueue_handle structure.  This further allows data path
operations such as push and pop to be inlined in hwqueue.h.

hwqueue: add scatter/gather support to hwqueue

Prior to this patch, hwqueue_push() assumed the "data" parameter was
the address of a buffer that needed to be mapped before being pushed.
Likewise, hwqueue_pop() assumed the value popped was the address of a
buffer that needed to be unmapped.

This patch set breaks out these operations into separate functions.
If the value to be pushed is the address of a buffer, it must be mapped
using hwqueue_map(), and the DMA address returned is then passed to
hwqueue_push(). Similarly, if the value retrieved by hwqueue_pop() is
the address of a buffer, it must be unmapped using hwqueue_unmap()
to get the virtual address of the buffer.

These changes BREAK the old semantics of hwqueue_push() and hwqueue_pop().
Expect compile errors due to argument list changes. Do NOT just cast the
old arguments to the new types; fix the errors properly.

hwqueue fix

Signed-off-by: Sandeep Paulraj <s-paulraj@ti.com>

hwqueue: clean up statistics to separate out error counters

This patch cleans up hwqueue statistics.  Specifically, it separates out
error counters from the original push/pop counters.

hwqueue: add device tree bindings documentation

Signed-off-by: Sandeep Paulraj <s-paulraj@ti.com>

hwqueue: move hwqueue doc

Signed-off-by: Sandeep Paulraj <s-paulraj@ti.com>