[RFC] dma-mapping: fix dma_mmap_user() for cma mem
When dma_alloc_coherent() returns memory allocated from CMA, we do not create
arm_vmregion maps for it.  Subsequent attempts to dma_mmap_coherent() this
memory fail to find an arm_vmregion map, and therefore fail to map the memory
to user.

This patch may not be strictly correct, but appears to work for me :-)