net: add device tree support for DaVinci MDIO
This patch adds device tree support in the DaVinci MDIO driver.
The davinci mdio driver has a way to set the bus frequency
using platform data. There is however no way to pass this information
from device tree.
This driver is currently being used in DaVinci and OMAP based SOCs which
do not yet have device tree support.
This patch first looks for a device tree node; if it finds
a node then it looks for the bus frequency from the device tree
bindings. Else it tries to obtain this from platform data.

The new binding is also being explained in the the Documentation for
davini_mdio

Signed-off-by: Sandeep Paulraj <s-paulraj@ti.com>