tci6614: Updating the cpgmac and mdio clock definition
Although the tci6614 spec clearly demarcates power domain 7 and 8; based
on experiments, there defintely is a dependency between the two domains.
Power domain 8 has to be enabled only after domain 7.
To this effect, the parent of the the cpgmac
and mdio clocks are being appropriately changed.

Signed-off-by: Sandeep Paulraj <s-paulraj@ti.com>

tci6614: add clock for mdio

As of now it is not possible to use clk_add_alias because the API
requires a "device" to be passed in as an argument.
Since this device is being defined in the clock lookups, there
is currently no way to get this information.

This commit essentially achieves something we would have achieved using
the clock_add_alias

Signed-off-by: Sandeep Paulraj <s-paulraj@ti.com>

tci6614: adding additional ethernet related clocks

This patch adds additional clocks for ethernet related clocks

Signed-off-by: Sandeep Paulraj <s-paulraj@ti.com>