tci6614: add clock for mdio
As of now it is not possible to use clk_add_alias because the API
requires a "device" to be passed in as an argument.
Since this device is being defined in the clock lookups, there
is currently no way to get this information.

This commit essentially achieves something we would have achieved using
the clock_add_alias

Signed-off-by: Sandeep Paulraj <s-paulraj@ti.com>