hwqueue: avoid divide on queue push/pop
This patch speeds up hwqueue push and pop operations by eliminating a divide
operation in the processing.  Since the ARM CPU does not have a divide
instruction, this operation was spending measurable time in software divide
routines.

This divide was being done in mapping an hwqueue instance pointer to and from
a queue index.  With this patch in place, we round up the size of the hwqueue
instance data to the nearest power of two, and use faster shift operations
instead.