dma: keystone: rename dma data and transfer direction fields
This patch simply renames dma_data_direction and dma_transfer_direction fields
to improve readability and reduce confusion while reading through the code.

The dma_data_direction is used for dma map/unmap APIs, i.e., as inputs to the
DMA-APIs.  The dma_transfer_direction, on the other hand, is an argument to
the DMA Engine APIs.