dma: keystone: add lock free descriptor accounting
This patch modifies the descriptor accounting code in the DMA implementation
to use lock-free techniques instead.  This way, we get by without having to
protect list maintenance operations with locks.

The lock-free list (actually more of a stack than a list) implementation is
index based, i.e., it maintains a singly linked list using 16-bit descriptor
indices instead of pointers.  In addition to the descriptor index, it
maintains a sequentially incremented token as some sort of a generation
counter.  This generation counter is used to solve the ABA problem inherent
with cmpxchg based lock-free singly linked lists (see [1] for details).

[1] http://en.wikipedia.org/wiki/ABA_problem