dma: keystone: remove descriptor state tracking
The original implementation of the packet dma driver had paranoid atomic state
transitions on a per-descriptor basis.  This was implemented at the time to
catch bugs in the implementation.  Now that we are past that stage, we no
longer need such detailed state tracking, and so this patch cleans this up.