ASoC: SAMSUNG: Add I2S soundcard using stub codec
HDMI can be configured to source audio data from I2S bus. For boards
on which there are no external audio chips, we still need to create
a I2S-based soundcard with a dummy codec driver to enable I2S subsystem
to pump data to the I2S bus, which can be used by HDMI module.

Signed-off-by: Tushar Behera <tushar.behera@linaro.org>