hwqueue: keystone: add accumulator support
The accumulator is a piece of PDSP firmware that pops descriptors from a queue
and collects them in a memory based ping-pong buffer.  In addition to this
functionality, it provides a configurable interrupt pacing capability that can
be used in some instances to reduce the interrupt rate on high packet rate
channels.

This patch adds support for both single-queue and multi-queue variants of
accumulator channels.  Multi-queue channels are capable of monitoring up to 32
queues in a single channel.  Both variants have been verified by rewiring the
ethernet receive queue through the accumulator.

This patch is nowhere near final, the following work items remain to be done:

1. Performance problems - for some reason the ethernet throughput appears to
   be terrible when running on an accumulator channel.  This is true for both
   single-queue and multi-queue cases, and needs further investigation.

2. Device tree bindings - this patch extends the keystone hwqueue device tree
   bindings, but does not document the extensions as is the norm.  This will
   need to be addressed in a future version of this patch.

3. Hardware spinlock support - the accumulator channel programming process is
   not atomic.  Therefore, when running in a multi-core system shared with
   DSPs, we need to acquire a hardware spinlock to protect the channel
   programming sequence.  This capability does not exist in this patch, and
   will need to addressed in a future verison.

4. List overruns - with this patch, accumulated queues maintain a page long
   list of descriptors in software.  This list is maintained using atomic
   operations, and is fed from the ping-pong list in the interrupt handler.
   There is a possibility for this list to get overrun, and this case is not
   handled in the code.

5. PDSP ID Handling - The device tree PDSP ID mapping between accumulator
   channels and PDSPs is obscene.  This should be replaced with phandles or at
   least PDSP names instead of having a convoluted ID thingy.

6. Termination and cleanup - This driver in general (not restricted to this
   patch) has lousy cleanup and termination handling.  This needs to be
   revisited.

Credits for the original patch go to Hao Zhang <hzhang@ti.com>.

Signed-off-by: Cyril Chemparathy <cyril@ti.com>

hwqueue: fix flush on accumulator channels

This patch fixes the handling of queue flush on accumulator channels

hwqueue: keystone: fix bug in the irq free routine

There was a small calculation mistake while calculating the ir number
of a queue that needs to be freed.
This patch fixes the issues

Signed-off-by: Sandeep Paulraj <s-paulraj@ti.com>

hwqueue: keystone: fix accumulator interrupt re-triggering

The accumulator interrupt re-triggering mechanism in place earlier was flawed.
Writing a set bit to the status register internally bumps up the pending IRQ
count on the QMSS interrupt distributor.  This means that we cannot rely on
the count == 0 condition in the ISR to distinguish between normal list flips
and re-triggered IRQs.

This patch adds explicit re-trigger accounting to circumvent this problem.
This is temporary at best.

hwqueue: keystone: update accumulator driver code for new command interface

Signed-off-by: Sandeep Paulraj <s-paulraj@ti.com>

hwqueue: keystone: add range ops support in accumulator