net:keystone: add ethernet subsystem driver
This commit adds an ethernet subsystem driver for the switch on
keystone devices. It initializes the host port and initializes
the slave ports. the number of slave ports to be initialized depends
on the device tree bindings that are deciede by the user. Depending
on the device tree bindings, the appropriate sgmii api's will
be called to setup the sgmii link in an appropriate way.

If there is a phy, then the driver will call phylib api's to monitor
the phy.

Signed-off-by: Sandeep Paulraj <s-paulraj@ti.com>