dma: keystone: add support for more than 4096 queues
The existing implementation of the keystone dma engine prevented queues
above 4095 from being used. This commit adds support in the keystone dma
engine to enable queues greater than 4095 to be used.
Any queue from 4096 to 8192 logically belongs to another queue manager.
So based on the queue number we calculate the queue manager number.
The existing implementation was never writing the queue manager number into
the appropriate locations in the descriptor and the packet dma rx flow
register. At reset this value was zero and things worked. While using any
queue above 4095, we have to appropriatley calculate the queue manager number
and put it in the descriptor and the appropriate packet dma rx flow register.

Also, the driver was not modifying the qm_base_address registers in the
global packet dma control registers. Again on reset, the appropriate value is
set into QM0 base address register and the others are 0. This also prevents
queues greater than 4095 from being used. To this effect, device tree support
is being added. We can have a maximum of 4 logical queue managers but in the
current generation of devices we have 2. From device tree we apprise the
driver of the number of queue managers using the binding
logical-queue-managers. In the device tree we also define the
qm-base-address for each logical queue manager using an array.

Signed-off-by: Sandeep Paulraj <s-paulraj@ti.com>