ARM: davinci - restructure header files for common clk migration
As part of migrating to common clk framework usage in DaVinci SoCs,
clock.h and clock.c are being deprecated as the pll and psc clock
drivers are moved to drivers/clk/davinci. The relevant PLL defines
are moved to include/mach/pll.h and will be used by SoC specific code
instead of using clock.h. The psc.c code is being deprecated as the
psc specific code is moved to the drivers/clk/davinci. So the extern
definitions are not used anymore and is now conditionally included.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>