net: keystone: NETCP Plug-in module support
This patch implements the changes in the NETCP core driver to support
plug-in module support. The biggest change is in the order of
initialization: prior to this patch, all sub-modules had to register
with the core driver BEFORE the core driver was probed. Since this
isn't possible if the sub-modules are loaded after the core driver,
the core driver now expects to initialize first and supports registration
and removal of sub-modules afterward.

The core now also supports multiple instances of NETCP hardware, and
multiple interfaces for each instances.

Here's a brief description of calls a plug-in module must support and
when they're called:

 1) probe() -- Called once per NETCP instance when the module is registed.
		The module should allocate a per-instance data structure
		and pass its address back via the inst_priv parameter.
 2) attach() -- Called when a network interface is created. The module
		should allocate a per-interface data structure and pass its
		address back via the intf_priv parameter.
 3) release() -- Called when a network interface is destroyed. The module
		should free its per-interface data structure here.
 4) remove() -- Called once per NETCP instance when the module is unregisted.
		The module should free its per-instance data structure here.

Known deficiencies include a lack of locking on the various linked lists,
failure to manage module reference counts, and inability to handle modules
installed while interfaces are up. There is also a known crash if the core
module is removed and reinstalled. These need to be addressed in future
patches.

Here is a summary of the changes to public data structure definitions
used by all components of the Keystone NETCP drivers:

  * Completely remove the "netcp_module_data" structure. The function
    entry vectors it previously contained are now part of the
    "netcp_module" structure.
  * Add a "netcp_device" structure. This describes a single instance
    of the NETCP hardware module, as a SoC could have more than one.
  * Change the parameters to the function entry vectors. Since there
    may be multiple instances of the NETCP hardware, and each instance
    may support multiple interfaces, functions may now take pointers
    to the "netcp_device" structure, a "inst_priv" (a plug-in defined,
    per-instance, private value), and/or a "intf_priv" (a plug-in
    defined, per-interface, private value).
  * Added "attach" and "release" function vectors. These functions are
    used by a plug-in module to get notified when a new interface is
    created or deleted.

---------

This patch includes minimal changes in ETHSS to support the API
changes. This module is now responsible for requesting the creation
and removal of network interfaces, as it is the driver that has to
manage the low-level interface configuration.

The intent is for this module to plug-in to the NETCP Core driver,
but is currently compiled into the same loadable module.

Support for multiple NETCP instances and multiple interfaces per
instance will be implemented in a future patch.

---------

This patch implements minimum changes to the PA to support the new API.
The only functional change is that the control of the Streaming
Switch has been moved to the NETCP Core driver. This allows the
ETHSS to properly initialize the switch to allow operation
without use of the PA, while also allowing this driver to re-
direct packets to the PA once it is configured.

Multiple instance and interface support will be added in a
future patch.

---------

This patch also implements minimal changes to the SA to conform to
the new API.