dma: keystone: Fix bug in scatterlist transmit
When a scatterlist containing multiple data buffers is passed
to chan_prep_slave_sg(), it was not setting the Packet Length
field of the first descriptor properly. This field should be
set to the total length of the packet, but was being set to
the size of the first data buffer. This caused all remaining
buffers to be ignored.

This fix accumulates the total packet size while assembling
the descriptor chain, so that the first descriptor gets set
appropriately. Note that the second and following descriptors
will have their packet length fields set to bogus values, but
these are ignored.