hwqueue: keystone: use extended hwqueue push and pop api's
The keystone hardware queue push api is being updated to use the
new generic hardware queue push api. In this an argument called flags
is passed into the push api. If bit 31 is set, the lower 17 bits will
have the packet size. This packet size is then written to the
packet size register.
An extra flags argument is also passed into the keystone hardware
queue push api though at present this has no use as of now.

Also writes to the packet size and ptr_size_thresh registers are now
protected with a lock.

Signed-off-by: Sandeep Paulraj <s-paulraj@ti.com>