net: keystone: Fix possible failure to poll for TX completions
If the NETCP core queues more than dma_poll_threshold packets to
a DMA channel before any packets arrive on the completion queue,
it will "forget" to poll until it runs out of free descriptors.
This is because it only called dma_poll() when dma_poll_count
became zero. This was done to avoid having multiple threads calling
dma_poll() on the same channel simultaneously.

This patch changes the criteria for polling so one is needed anytime
dma_poll_count is less than or equal to zero. Since this opens the
possibility for multiple threads to call dma_poll(), a spinlock is
added to the tx_pipe structure to act as a mutex. Note that if one
thread is already in dma_poll() we don't need to start another, so
it NEVER spins waiting for this lock.