tci6614: dt: add OF irq domain and chip irq bindings
Define a virtintc controller irq domain and chip irq DT bindings
for appleton. All of the omap AINTC, CPINTC and CPINTD controllers
are combined into the virtintc controller domain. Use interrupts property
to refers to the offset of the device irq in the domain.
The interrupt-cells defines only one cell that has the device irq number
used by request_irq().

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>