tci6614: dt: change interrupt cells to 2 to include irq flags
of driver uses the second cells to read the irq flag value from the DT
bindings and set the irq type. So change the virtintc irq controller
bindings to reflect this and modify all device bindings to use 2 cells
for interrupts attribute, first indicating irq number and second the
flag.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>