irqchip: add tci6614 complex irqchip driver
This patch consolidates TCI6614's bundle of interrupt controllers (CPINTC,
CPINTD, and AINTC) in a single driver outside mach-davinci code.

This is a step towards migrating away from mach-davinci altogether, but is not
expected to be posted upstream anytime soon.

If this code is to be posted upstream, we'll need to cleanup the domain setup,
and move the chained mappings to DT at the very minimum.  Ideally, we'd need
to split these up into separate drivers irqchip drivers for each block.

irqchip: tci6614 updates and cleanups