clocksource: add support for tci6614 timers
This patch adds a slightly simplified variant of the davinci timer64 code
under drivers/clocksource.  This is a _temporary_ solution in preparation for
moving appleton support out of mach-davinci and into mach-keystone.  Ideally,
we'd need to unify this driver along with the c6x timer64 driver and the
davinci timer driver, but that's another battle for another day.

clocksource: tci6614 updates and cleanups