gpio: keystone: add gpio driver
This patch adds GPIO driver for keystone devices. These devices
have a simple GPIO module that can have upto 4 pair of banks. Each
bank supports upto 16 pins.  Also adds irq controller support. This
registers a irq domain for the gpios and maps the same to the
host side irqs. Users of gpio will be able to set irq parent to the
gpio irq controller in the device bindings and set the interrupts
property to configure the GPIO interrupt line for use.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>