ARM: perf: [WIP] Initial bL switcher support
This patch adds preliminary, highly experimental perf support for
CONFIG_BL_SWITCHER=y.

In this configuration, every PMU is registered as valid for every
logical CPU, in a way which covers all the combinations which will
be seen at runtime, regardless of whether the switcher is enabled
or not.

Tracking of which PMUs are active at a given point in time is
delegated to the lower-level abstractions in perf_event_v7.c.

Warning: this patch does not handle PMU interrupt affinities
correctly.
Because of the way the switcher pairs up CPUs, this
does not cause a problem when the switcher is active; however,
interrupts may be directed to the wrong CPU when the switcher is
disabled.
This will result in spurious interrupts and wrong event
counts.

Signed-off-by: Dave Martin <dave.martin@linaro.org>