ARM: perf: [WIP] Manipulate the right shadow register for PM*CLR
Where the ARM Architecture exposes PM*SET and PM*CLR, these really
manipulate the same underlying register.

This patch uses the PM*SET register for storing the logical state
when the PMU is not active, and mainpulates that state when the
code attempts to access the corresponding PM*CLR register.

PMOVSR is a special case: this is a reset-only register, so the
logical copy of PMOVSR is always used.

These changes result a small number of unused fields in the
armv7_pmu_logical_state structre.  For now, this is considered to
be harmless -- it may be tidied up later.

Signed-off-by: Dave Martin <dave.martin@linaro.org>