ARM: perf: [WIP] Add register emulation for offline ARMv7 PMUs
This patch aims to provide basic register file functionality for
ARMv7 CPU PMUs while the PMU is offline.
It is incomplete and
lacks the necessary plumbing to actually make use of this, but the
extra code needed is not expected to be large or complex.

Save/restore are ported over the register emulation framework,
since the offline logical state of the CPU matches exactly what
needs to be captures in save/restore.

Because this patch is rather invasive, it should be
dropped in the future in favour of higher-level abstraction before merging upstream.

Signed-off-by: Dave Martin <dave.martin@linaro.org>