ARM: perf: [WIP] Check PMU is valid for the CPU in armpmu_disable()
This hack is required in order to be able to manipulate the CPU
logical state safely in the absence of the b.L switcher, for test
purposes.

The other similar checks are already present in the b.L MP perf
support patches.

Normally, only the physical PMU state should be mainpulated in a
kernel which doesn't include the switcher, so it may be possible to
remove this patch later.

Signed-off-by: Dave Martin <dave.martin@linaro.org>