clk: keystone: enhancement to support ARMPLL, DDRPLL and PAPLL
The clk driver for Main PLL is enhanced to support additional PLLs
found on keystone2 devices. The addtional PLLs are controlled by
the register where as the Main PLL is controlled by a PLL
controller. Add a has_pll_cntrl boolean to indicate if the PLL
is controlled by a pllctrl or not. Some updates made to make code
more generic.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Sandeep Nair <sandeep_n@ti.com>