net: keystone: Fix bug in TCP/UDP Software checksum offload
The software checksum offload for TCP/UDP packets was broken in the
case where (1) there are multiple fragments, and (2) a fragment other
than the last contained an odd number of bytes.

This bug was introduced by attempting to have a single function that
would walk an SKB tree and apply either the TCP/UDP checksum or SCTP
CRC32C algorithm. This patch uses the standard skb_checksum_help()
function for TCP/UDP packets, and creates a new (local) function
skb_sctp_csum_help() to generate SCTP CRC32Cs.

In the process of diagnosing this bug, it was noticed that the original
packet was being modified in-place. Sometimes this is NOT a good thing,
and this patch addresses this case by cloning the packet header. The
only known situation where this occurs is when a transmit timestamp has
been requested, so the impact of cloning the header is low.

Signed-off-by: Reece Pollack <x0183204@ti.com>

Conflicts:

	drivers/net/ethernet/ti/keystone_pa.c