ARM: perf: [WIP] Map PMU IRQ affinities correctly
This patch determines the (mpidr,irq) pairings associated with each
PMU from the DT at probe time, and uses this information to work
out which IRQs to request for which logical CPU when enabling an
event on a PMU.

This patch also ensures that each PMU's init function is called on
a CPU of the correct type.
Previously, this was relying on luck.

Signed-off-by: Dave Martin <dave.martin@linaro.org>