ARM: EXYNOS: Set arch_sys_counter as default clocksource
The Exynos 5250 includes an architected timer which is on core and
very cheap to read.

Enabling the architected timer can lead to a significant performance
improvement for timer sensitive workloads such as TCP/IP.

It turns out that both arch_sys_counter and mct-frc have a rating of
400. On bootup, arch_sys_counter is initialized first then the
clocksource changes to mct-frc when that initialises later on.

If the rating of mct-frc is reduced below 400, then arch_sys_counter
becomes the default.

Suggested-by: Steve Capper <steve.capper@linaro.org>
Signed-off-by: Tushar Behera <tushar.behera@linaro.org>