net: keystone: enable mac-phy mode with no mdio
This commit adds a new mode where there is a phy on the evm
that is always enabled and on.
For this we are defining a new mode called
SGMII_LINK_MAC_PHY_NO_MDIO. In this case, mdio will not be used to
communicate with the phy. The device tree documentation
is also being appropriately updated

Signed-off-by: Sandeep Paulraj <s-paulraj@ti.com>