net: keystone: update ioctl handling in net core engine
This commit updates the module ioctl handling. Error codes in the form of
negative numbers are treated as failures and the negative error code is
appropriately returned. The -EOPNOTSUPP is however treated in a slightly
different manner. We don't treat this as a failure and use this value to
continue calling out the ioctl callouts for other modules.
If a zero is returned from a module ioctl callout we store this value.
We used this stored value as a measure of success and return a "0"
even if one ioctl callout returns a zero and the others return a -EOPNOTSUPP.
Also in this commit, the skb_tx_timestamp is being called at its
appropriate place.

Signed-off-by: Sandeep Paulraj <s-paulraj@ti.com>