net: keystone: update 10g switch ioctl handling
The ioctl handling in the keystone network infrastructure is being
updated and knowledge of whether the pa timestamp is required or not
is not supposed to be in any module apart from packet accelerator.
To this effect the timestamp and ioctl handling is being updated in the
switch driver. The 10g switch driver is a clone of the 1g switch driver
and thus needs to updated as well though the 10g data path does not have
a packet accelerator.

Signed-off-by: Sandeep Paulraj <s-paulraj@ti.com>