hwqueue: keystone: add multi tree support
The keystone qos driver has support for debugfs/sysfs.
The current version of the driver uses "qos" as the top directory
name for both debugfs/sysfs.
This will potentially create issues when we have multiple trees.
This commit adds support for adding a directory name based on
the node name/instance.

Signed-off-by: Sandeep Paulraj <s-paulraj@ti.com>