dma: keystone: Fix crash when receiving DMA fragments
The underlying bug is an incomplete replacement of virtual/dma address
conversions in drivers/dma/keystone-pktdma.c:chan_complete()

The Appleton kernel (3.3) used phys_to_page() while filling in the scatterlist
entries. This doesn't work on the Keystone due to the weird address mappings.
There isn't a direct replacement for phys_to_page(), so this patch creates a
macro called dma_to_page(), which invokes the proper series of lesser functions.

Unfortunately, the dma_to_pfn() function wants a pointer to a "struct device",
probably so it can apply the proper address translation, but the struct device
used for the page_to_phys() mapping used a higher layer's struct that we don't
have access to. Currently this parameter isn't used, so we can get away with
providing a bogus address. NULL is specifically chosen here so that if this
parameter starts to be used we'll get an immediate crash, rather than some non-
obvious behavior.

Yes, it's a hack. Yes, it should be done a different way. If you have a
constructive suggestion, make it known.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>