i2c: s3c2410 : Add polling mode support
This patch adds polling mode support for i2c-s3c2410 driver.
The I2C_SATAPHY controller in exynos5250 lacks an interrupt
line and the i2c-s3c2410 driver is only interrupt driven, thus
a polling mode support is required in the driver for supporting
this controller. This patch adds this support to the driver.
The SATAPHY bus is an internal bus with I2C controller as master
and SATA PHY controller as the slave.

Signed-off-by: Vasanth Ananthan <vasanth.a@samsung.com>
Signed-off-by: Tushar Behera <tushar.behera@linaro.org>