tci6638: dts: Support 6 TX queues for QoS
Current sample QoS trees require 6 TX queues. For some reason this was
configured for only 4 queues. This patch changes the value of intf_tx_queues
to correct that.

Also, there were per-interface entries for intf_tx_queues, but these
are ignored. This patch removes these to avoid further confusion.

Finally, set intf_tx_queues to 1 for the XGE devices, as QoS is not
configured on them.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>